requirements.txt
-i https://pypi.org/simple
asgiref>=3.6.0,<4
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
deprecated==1.2.13
diff-match-patch==20200713
django==4.2
django-admin-autocomplete-filter==0.7.1
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.8.5
django-admin-sortable==2.3
django-appconf==1.0.5
django-apptemplates==1.5
django-auth-ldap==4.1.0
django-autocomplete-light==3.9.4
django-configurations==2.5 
django-cors-headers==3.13.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-dirtyfields==1.8.2
django-downloadview==2.3.0
django-filter==22.1
django-gravatar2==1.4.4
django-ical==1.8.3
django-import-export==3.0.0b4
django-logentry-admin==1.1.0
django-model-utils==4.2.0
django-oauth-toolkit==2.2.0
django-phonenumber-field==6.3.0
django-polymorphic==3.1.0
django-recurrence==1.11.1
django-registration==3.3
django-rest-assured==0.2.3
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-select2==7.10.0
django-settings-export==1.2.1
django-silk==5.0.1
django-tables2==2.4.1
django-wkhtmltopdf==3.4.0
djangorestframework==3.14
djangorestframework-filters==1.0.0.dev2
et-xmlfile==1.1.0
factory-boy==3.2.1
faker==13.15.0
gprof2dot==2021.2.21
gunicorn==20.1.0
humanize==4.2.3
icalendar==5.0.0a1
idna==3.3
importlib-metadata==4.12.0
itypes==1.2.0
jinja2==3.1.2
jwcrypto==1.3.1
markdown2==2.4.3
markuppy==1.14
markupsafe==2.1.1
mysqlclient==2.1.1
mysql-connector-python==8.0.33
django-mysql==3.9.0
oauthlib==3.2.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.10
phonenumberslite==8.12.52
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pygments==2.12.0
pypugjs==5.9.11
python-dateutil==2.8.2
python-ldap==3.4.2
python-redmine==2.3.0
pytz==2022.1
pyyaml==6.0
requests==2.28.1
setuptools==67; python_version >= '3.7'
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
whitenoise==6.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
